+++ src/EuclideanSpace/Point3.Extensions.cs
@@ -1,0 +1,44 @@
+namespace EuclideanSpace
+{
+    using System.Diagnostics;
+    using System.Numerics;
+    using System.Runtime.CompilerServices;
+
+    public static partial class Point3
+    {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static Vector3<TScalar> AsVector3<TScalar>(this Point3<TScalar> point)
+            where TScalar : IAdditionOperators<TScalar, TScalar, TScalar>,
+            IMultiplyOperators<TScalar, TScalar, TScalar>,
+            ISubtractionOperators<TScalar, TScalar, TScalar>,
+            IUnaryNegationOperators<TScalar, TScalar>,
+            IDivisionOperators<TScalar, TScalar, TScalar>
+            => new(point.X, point.Y, point.Z);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static TScalar GetElementUnsafe<TScalar>(in this Point3<TScalar> point, int index)
+            where TScalar : IAdditionOperators<TScalar, TScalar, TScalar>,
+            IMultiplyOperators<TScalar, TScalar, TScalar>,
+            ISubtractionOperators<TScalar, TScalar, TScalar>,
+            IUnaryNegationOperators<TScalar, TScalar>,
+            IDivisionOperators<TScalar, TScalar, TScalar>
+        {
+            Debug.Assert(index is >= 0 and < Count);
+            ref var address = ref Unsafe.AsRef(in point._x);
+            return Unsafe.Add(ref address, index);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static TScalar GetElement<TScalar>(this Point3<TScalar> point, int index)
+            where TScalar : IAdditionOperators<TScalar, TScalar, TScalar>,
+            IMultiplyOperators<TScalar, TScalar, TScalar>,
+            ISubtractionOperators<TScalar, TScalar, TScalar>,
+            IUnaryNegationOperators<TScalar, TScalar>,
+            IDivisionOperators<TScalar, TScalar, TScalar>
+        {
+            ThrowHelpers.ThrowIfGreaterThanOrEqual((uint)index, (uint)Count);
+
+            return point.GetElementUnsafe(index);
+        }
+    }
+}