+++ src/EuclideanSpace/Vector2_1.cs
@@ -58,6 +58,7 @@ namespace EuclideanSpace
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public TScalar LengthSquared() => Vector2.Dot(this, this);
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Vector2<TScalar> operator +(Vector2<TScalar> left, Vector2<TScalar> right) =>
             new(left.X + right.X, left.Y + right.Y);