+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.1.1-preview] - 2024-08-11
+
+### Added
+
+- `Vector3<TScalar>`, `Vector3`
+- `Vector3Conversions<TTarget>`, `Vector3Conversions`
+- `Point3<TScalar>`, `Point3`
+- `Point3Conversions<TTarget>`, `Point3Conversions`
+
 ## [0.1.0-preview] - 2024-08-07
 
 ### Added
@@ -9,6 +18,8 @@
 - `Point2<TScalar>`, `Point2`
 - `Point2Conversions<TTarget>`, `Point2Conversions`
 
-[Unreleased]: https://github.com/qbit86/euclidean-space/compare/euclidean-space-0.1.0-preview...HEAD
+[Unreleased]: https://github.com/qbit86/euclidean-space/compare/euclidean-space-0.1.1-preview...HEAD
+
+[0.1.1-preview]: https://github.com/qbit86/euclidean-space/compare/euclidean-space-0.1.0-preview...euclidean-space-0.1.1-preview
 
 [0.1.0-preview]: https://github.com/qbit86/euclidean-space/releases/tag/euclidean-space-0.1.0-preview